Move test entities to separate files
